Warn when importing DA.Generics
The module `DA.Generics` is hidden from docs of `daml-stdlib` because we
consider it experimental. On top of that, the comments at the module
make it clear that this module does work across SDKs. Nevertheless,
let's make it even clearer that `DA.Generics` does not work with
data-dependencies and will stop you from upgrading your package.

I've decided against using the a `{-# WARNING ... #-}` annotation
on the `DA.Generics` module declaration since that issues a warning for
every single time an enitity from the module is used rather than once
when importing the module. To me, that's a bit over the top since it is
conceivable that people still want to use the module during some stages
of their development. Another thing we might need to consider soon is a
`-fno-data-dependencies-warnings` flag which suppresses warning related
to data-dependencies. This would be impossible with a
`{-# WARNING ... #-}` annotation.

CHANGELOG_BEGIN
[DAML Stdlib] Issue a warning when importing the `DA.Generics` module,
which does not work with data-dependencies.
CHANGELOG_END