Adapt damlc integration tests to newest ghcide
We've recently upgraded to a newer ghcide version in #7706 which fixes
the bug that column numbers in diagnostics printed on the command line
were off by one.

This PR brings the ranges in our test files in line with what the
command line tools print, which is also what vscode shows.

CHANGELOG_BEGIN
CHANGELOG_END