json-api: Aligning portfile behaviour
While the sandbox can be pointed to an existing port file to write its
port to, the http json api will fail if the port file already exists.
This is annoying in particular for testing. This PR aligns the
behaviours such that the http json api appends to existing port files
instead insisting on creating a new file.

CHANGELOG_BEGIN
CHANGELOG_END