work around Azure/GitHub disagreement
Azure used to report the status of the entire build to GitHub, which we
use as the "required check" for PRs to be merged. Ir doesn't do that
anymore which means we can't merge anything. It's unclear whether or not
that is a deliberate change.

This attempts to work around that by creating an extra job that depends
on all the other, which GitHub could depend on.

CHANGELOG_BEGIN
CHANGELOG_END