Parametrized performance envelope tests
Before this change, the performance envelope tests used a hard-coded
number of pings and warmup pings for the tests, in particular for the
throughput test. As a result, a performance test with 100tps was only
running 2 seconds. Now, the number is derived from the target throughput
numbers.

CHANGELOG_BEGIN
CHANGELOG_END