DALF Reader: Refactor decoder environment (#7891)
* DALF Reader: Refactor decoder environment

Currently, the module to be decoded is part of the decoder environment.
This turned out to be unpleasant during my attempts to implement
interning for types since decoding the type interning table does not
happen in the context of a module.

This PR moves the module out of the decoder environment and passes it
to the decoder function for modules directly. Unfornatunately, we still
need to keep the module name in the environment since that is used as a
default when decoding location information. We solve this problem by
making the module name in the environment optional and always filling
it in when decoding in actual module.

CHANGELOG_BEGIN
CHANGELOG_END

* Adapt tests

CHANGELOG_BEGIN
CHANGELOG_END

* Follow Remy's suggestion

CHANGELOG_BEGIN
CHANGELOG_END