hlint: Suggest fmap f . V.fromList over V.fromList . map f
I somehow had a hunch that calling `V.fromList . map f`, where `V` is
`Data.Vector`, uses more memory than `fmap f . V.fromList`. My
reasoning was that in the former, allocating the vector needs to force
the spine of the list in order to get its length. This would mean, we
have to allocate a new cons cell for each element in the original list
and GHC cannot eliminate via deforestation. In contrast, with
`fmap f . V.fromList` we only need to allocate a single slot in a new
vector rather than a full cons cell. Also, garbage collecting the
intermediate vector should be faster than garbage collecting an
intermediate list since with the vector we only need to follow one
pointer for each element of the original list, whereas with a list we
have to follow two pointers.

Experiments confirm my suspicion. I've run both of the above examples
wth `f = (+1)` as well as a simple `V.fromList` on a list with
1,000,000 elements. The total allocated memory was as follows:

* `V.fromList` - 97MB
* `fmap f . V.fromList` - 137MB
* `V.fromList . map f` - 201MB

In other words, in `fmap f . V.fromList` mapping a single element takes
40 bytes, whereas in `V.fromList . map f` it takes 104 bytes. I would
call that difference significant.

Having learned that, I thought it would be a good idea to encode this
knowledge in an `hlint` rule.

CHANGELOG_BEGIN
CHANGELOG_END