Rewrite multiple binding lets sequentially. (#8006)
* Rewrite multiple binding lets sequentially.

This PR enforces a sequential evaluation and dependency on
multiple-binding lets, by rewriting,

  let x1 = e1
      x2 = e2
      ...
      xn = en
  in b

Into a chain of single-binding let expressions,

  let x1 = e1 in
    let x2 = e2 in
      ...
        let xn = en in
          b

And likewise, rewriting multi-binding let statements in a "do" block,

  do
      ...
      let x1 = e1
          x2 = e2
          ...
          xn = en
      ...

Into single-binding let statements,

  do
      ...
      let x1 = e1
      let x2 = e2
      ...
      let xn = en
      ...

This PR fixes #6788 and adds a bunch of tests.

This is a bugfix, but it may break existing code that relies on
out-of-order let bindings. The suggested fix is to simply reorder
the let bindings, placing them in order of their dependency.

CHANGELOG_BEGIN

- [DAML Compiler] The DAML compiler now enforces sequential
  ordering for ``let`` expressions with multiple bindings.
  This fixes a bug where ``let`` expressions would be evaluated
  in the wrong order.

CHANGELOG_END

* Update compiler/damlc/daml-preprocessor/src/DA/Daml/Preprocessor.hs

Co-authored-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>

* adjust last binding behavior

Co-authored-by: Moritz Kiefer <moritz.kiefer@purelyfunctional.org>