Sandbox: In tests, wait for Sandbox Classic to "properly" start.
There's a potential race condition in tests, where we could submit a
request _before_ the ledger has processed its own configuration. This
results in an `UNAVAILABLE` gRPC error: "The ledger configuration is not
available."

This changes the `SandboxFixture` abstract class to wait for the server
to intialize its config before considering it to be properly "started".

CHANGELOG_BEGIN
CHANGELOG_END