Remove password option from Navigator
This was leftover from ancient times. You couldn’t actually use this
in any way since we always defaulted to the select method and provided
no way to change this. We still support it in the config but emit a
warning now if you use it.

changelog_begin

- [Navigator] The `password` option in the Navigator config file is
  now deprecated. Note that it was already unused before.

changelog_end