Mark NPM snapshots with --tag next
NPM specialcases the latest tag which is used by default and defaults
to that in things like `npm add`. Given that we don’t want people to
rely on snapshots, tagging them as next (common convention on npm)
seems like a good idea.

I have no idea how to test this reasonably so the next snapshot will
have to tell if it works or not.

changelog_begin
changelog_end