Add daml.yaml in temp directory during SDK install
By adding a daml.yaml in the temp directory, we prevent issues that
could arise from a spurious daml.yaml existing in parent directories
(e.g. in /tmp). The daml.yaml contents are not important for the
installation, but if there's an error while reading daml.yaml, the
installation would fail.

This fixes #5720 in a backward-compatible way. I tested the fix manually.
I'm not sure there's an easy way to automate the test.