separate CI config for master & PRs
Our CI configuration is becoming unwieldy; this PRs attempt to get back
some measure of control by splitting the CI configuration of the master
branch and the CI configuration of PRs, as there are a number of jobs
that should run only on one of them.

Because there is a bit of setup involved (at the Azure level), this PR
does not attempt to remove all duplication; that will be done as future
PRs.

I have kept the changes in separate commits for ease of reviewing: the
first commit just copies the entire `azure-piplines.yml` file to
`ci/prs.yml`, and the second commit prunes from each what they should
never run. Reviewers should focus on the second commit.

CHANGELOG_BEGIN
CHANGELOG_END